third_party/WebKit/LayoutTests/http/tests/security/subresourceIntegrity/subresource-integrity-hash-function-priority-console-messages-expected.txt

CONSOLE ERROR: Failed to find a valid digest in the 'integrity' attribute for resource 'http://127.0.0.1:8000/security/subresourceIntegrity/call-success.js' with computed SHA-256 integrity 'B0/62fJSJFrdjEFR9ba04m/D+LHQ+zG6PGcaR0Trpxg='. The resource has been blocked.
